requirements.txt
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.2
scikit-learn==1.5.1
scikit-optimize==0.10.2
imbalanced-learn==0.12.4
jupyter==1.1.1
seaborn==0.13.2
tensorflow==2.18.0
xgboost==2.1.1
termcolor==2.4.0
tabulate==0.9.0
InquirerPy==0.3.4
tensorflow==2.16.2
python-dotenv==1.0.1
pyarrow==18.0.0
joblib==1.4.2